Bugfixes to reproduce output of the original pipeline
A number of bugs was fixed in order to reproduce the output of Uli's
original script.

In particular, there was a missing exclusion of non-PASS variants, and
a wrong input file for the hmm probabilities calculation.

When tested on 2 F2 samples, the final output of one was the same as for
the original script, and the other was not the same but close.