Fix rule 03_prepare_marker_files.smk/qc_marker_snp_counts crash
The rule 03_prepare_marker_files.smk/qc_marker_snp_counts produced an
error when run on the MPI SGE cluster:

Traceback (most recent call last):
   File "tiger-pipeline/.snakemake/scripts/tmpavlb3k8w.plot_parental_biallelic_snp_counts.py", line 5, in <module>
     import allel
ModuleNotFoundError: No module named ‘allel’

When run locally, this error does not occur. Since the rule is not
resource-intensive, fix this by adding the rule to localrules. The
underlying cause of the error is still not clear, though.