Fix auto determination for parent ref from a/b
There was a bug in automatic selection of the ref parent from a/b
parents. The parent with the larger number of SNPs should be set to src,
the one with smaller number of SNPs to ref. The auto-selection was done,
but the steps to realize the subsequent processing were not complete,
and parent "a" was still used downstream. This commit fixes the issue.