Fix: add explicit .tbi dependency for f2 dbg variant calling
The problem occured most likely due to file latency issues on the
cluster, it could not be reproduced locally with the test data set.

The error message was

“ A USER ERROR has occurred: An index is required but was not found for file
drivingVariantFile:results/variants/f2/gvcf/Sha_ICE49.g.vcf.gz. Support
for unindexed block-compressed files has been temporarily disabled.
Try running IndexFeatureFile on the input"

The .tbi files are already produced, but not explicitely mentioned as
input/output in the rules, which might cause Snakemake not to wait until
the filesystem sees the files.